+++ stroom-core-client-widget/src/main/java/stroom/cell/clickable/client/HyperlinkType.java
@@ -1,6 +1,6 @@
 package stroom.cell.clickable.client;
 
-public enum HyperlinkTarget {
+public enum HyperlinkType {
     DIALOG,
     STROOM_TAB,
     BROWSER_TAB,