+++ stroom-core-client/src/main/java/stroom/users/client/UsersPlugin.java
@@ -5,7 +5,7 @@ import com.google.inject.Provider;
 import com.google.web.bindery.event.shared.EventBus;
 import stroom.alert.client.event.AlertEvent;
 import stroom.cell.clickable.client.Hyperlink;
-import stroom.cell.clickable.client.HyperlinkTarget;
+import stroom.cell.clickable.client.HyperlinkType;
 import stroom.core.client.ContentManager;
 import stroom.core.client.MenuKeys;
 import stroom.menubar.client.event.BeforeRevealMenubarEvent;
@@ -49,7 +49,7 @@ public class UsersPlugin extends NodeToolsPlugin {
                                 final Hyperlink hyperlink = new Hyperlink.HyperlinkBuilder()
                                         .title("Users")
                                         .href(usersUiUrl)
-                                        .target(HyperlinkTarget.STROOM_TAB)
+                                        .type(HyperlinkType.STROOM_TAB)
                                         .build();
                                 final IFrameContentPresenter presenter = presenterProvider.get();
                                 presenter.setHyperlink(hyperlink);